build: Add ability to override images with YAML
This commit adds a schema and parsing logic for an override.yml.
This file is applied to an already parsed Moby config and will replace
all instances of the source pattern with the substitute pattern.
This allows for easy replacement of images in a config file for
development or test.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>