Add `user-binds` to image config
These extra binds are appended (unconditionally) to the label mounts. Setting
`user-binds` in an image label is forbidden, i.e. they are only allowed in
assemblies.

This allows users to augment the set of binds specified in the image label with
some local ones. Previously to do this one would have to duplicate all the
binds from the image in the assemlby, which is error prone and likely to get
out of date.

There may be other fields where such a distinction would be valuable, this is
intended to validate the concept.

Signed-off-by: Ian Campbell <ijc@docker.com>