Fix #218 enforce all outgoing tar files to use PAX
While processing the content of a tar image, moby/tool was blindly
reusing the original tar format.
Moreover moby/tool locates the file under the prefix, so if the
original file was stored as USTAR, the filename length and new prefix
could be greater than the USTAR limit producing a fatal error.

The fix is to always enforce PAX format on all copied files from the
original image.

Signed-off-by: Brice Figureau <brice@daysofwonder.com>